refactor(mobile): route tracking
